Simplify impl modules
