--- serde_arrow/src/arrow2_impl/type_support.rs
@@ -1,7 +1,0 @@
-use crate::internal::error::Error;
-
-impl From<crate::_impl::arrow2::error::Error> for Error {
-    fn from(err: crate::_impl::arrow2::error::Error) -> Error {
-        Self::custom_from(format!("arrow2::Error: {err}"), err)
-    }
-}
